+++ packages/cardano-services-client/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.3](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services-client@0.17.2...@cardano-sdk/cardano-services-client@0.17.3) (2024-02-02)
+
+**Note:** Version bump only for package @cardano-sdk/cardano-services-client
+
 ## [0.17.2](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/cardano-services-client@0.17.1...@cardano-sdk/cardano-services-client@0.17.2) (2024-02-02)
 
 **Note:** Version bump only for package @cardano-sdk/cardano-services-client