+++ packages/hardware-trezor/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.10](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-trezor@0.4.9...@cardano-sdk/hardware-trezor@0.4.10) (2024-02-02)
+
+**Note:** Version bump only for package @cardano-sdk/hardware-trezor
+
 ## [0.4.9](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/hardware-trezor@0.4.8...@cardano-sdk/hardware-trezor@0.4.9) (2024-02-02)
 
 **Note:** Version bump only for package @cardano-sdk/hardware-trezor