+++ packages/input-selection/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.17](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.16...@cardano-sdk/input-selection@0.12.17) (2024-02-02)
+
+**Note:** Version bump only for package @cardano-sdk/input-selection
+
 ## [0.12.16](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.15...@cardano-sdk/input-selection@0.12.16) (2024-02-02)
 
 **Note:** Version bump only for package @cardano-sdk/input-selection